refactored locationControls to have component indexing logic
