+++ src/components/controls/locationComponents/LocationIndexItem.jsx
@@ -1,0 +1,13 @@
+import React from 'react'
+
+const LocationIndexItem = props => {
+  const {  } = props
+
+  return (
+    <strong>HERHEHREHHEE</strong>
+  )
+}
+
+
+export default LocationIndexItem
+
