+++ src/index.jsx
@@ -2,6 +2,7 @@ import React from 'react'
 import ReactDOM from 'react-dom'
 
 import Root from './components/root'
+import './stylesheets/header.css'
 import './stylesheets/controls.css'
 import './stylesheets/player.css'
 import './stylesheets/root.css'
